global-perspectives-starter/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
aiohttp==3.9.5
python-dotenv==1.0.1
rapidfuzz==3.9.6
spacy==3.7.5
geopy==2.4.1
# AWS Bedrock integration
boto3==1.35.0
# For tests
pytest==8.3.3
